系统管理/试卷管理   --完成编辑试卷接口

+++ backend/src/main/java/com/achu/aijumptest/mapper/PaperQuestionMapper.java
@@ -18,5 +18,5 @@ import java.util.List;
 @Mapper
 public interface PaperQuestionMapper extends BaseMapper<PaperQuestion> {
 
-    List<QuestionDetailVO> selectQuestions(@Param("id") Long id);
+    List<QuestionDetailVO> selectQuestions(@Param("id") Integer id);
 }
